autocomplete and local-react to typescript
