+++ package.json
@@ -31,7 +31,7 @@
     "test:watch": "yarn build && vitest",
     "instant-meilisearch:test:watch": "vitest --project \"@meilisearch/instant-meilisearch\"",
     "autocomplete:test:watch": "vitest --project \"@meilisearch/autocomplete-client\"",
-    "test:types": "yarn build && tsc --noEmit",
+    "test:types": "turbo test:types",
     "version-packages": "changeset version && turbo version",
     "release": "yarn build && changeset publish"
   },
