+++ playgrounds/autocomplete/index.html
@@ -1,16 +1,13 @@
-<!DOCTYPE html>
-<html>
+<!doctype html>
+<html lang="en">
   <head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width,initial-scale=1.0" />
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Demo of Meilisearch + autocomplete.js</title>
-    <div id="autocomplete"></div>
   </head>
-
   <body>
     <div id="app"></div>
-    <script type="module" src="./main.js"></script>
-    <script type="module" src="./src/app.js"></script>
+    <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
