--- playgrounds/autocomplete/main.js
@@ -1,8 +1,0 @@
-import './style.css'
-
-// eslint-disable-next-line no-undef
-document.querySelector('#app').innerHTML = `
-  <div>
-    <div id="autocomplete"></div>
-  </div>
-`
