+++ playgrounds/autocomplete/package.json
@@ -3,12 +3,14 @@
   "version": "0.0.0",
   "private": true,
   "description": "Javascript playground for autocomplete",
+  "type": "module",
   "scripts": {
-    "predev": "node setup.mjs",
+    "predev": "node setup.js",
     "dev": "vite",
     "setup": "node setup.mjs",
     "test:e2e": "concurrently --kill-others -s first \"yarn dev\" \"cypress run --env playground=local\"",
-    "test:e2e:watch": "concurrently --kill-others -s first \"yarn dev\" \"cypress open  --env playground=local\""
+    "test:e2e:watch": "concurrently --kill-others -s first \"yarn dev\" \"cypress open  --env playground=local\"",
+    "test:types": "tsc"
   },
   "keywords": [
     "autocomplete"
