+++ playgrounds/autocomplete/setup.js
