+++ playgrounds/autocomplete/src/style.css
