+++ playgrounds/local-react/vite.config.ts
@@ -1,9 +1,8 @@
-/* eslint-disable @typescript-eslint/no-require-imports */
-const { defineConfig } = require('vite')
-const react = require('@vitejs/plugin-react')
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
 
 // https://vite.dev/config/
-module.exports = defineConfig({
+export default defineConfig({
   server: { port: 5174, strictPort: true },
   plugins: [react()],
 })
